Ensure entities still colliding before resolving collision
This collisionCheckAgain is required because as we resolve collisions by moving entities around, the initial collection of collision events may become outdated. Checking that the two entities are still colliding is not a perfect solution (if there even is one), but it should be good enough... We can revisit this and attempt other solutions down the line if needed.